requirements.txt
Django==5.1.1
djangorestframework==3.15.2
Pillow==10.4.0
celery==5.4.0
redis==5.1.1